chore: 调整 worker 目录结构
