+++ packages/utils/src/workers/extrude_polyline.ts
@@ -1,5 +1,5 @@
 import { vec2, vec3 } from 'gl-matrix';
-import { aProjectFlat } from '../../geo';
+import { aProjectFlat } from '../geo';
 const tmp = vec2.create();
 const capEnd = vec2.create();
 const lineA = vec2.create();
