+++ packages/utils/src/workers/polygonFillModel.ts
@@ -1,5 +1,5 @@
 import { IEncodeFeature, IVertexAttributeDescriptor } from '@antv/l7-core';
-import { encodePickingColor } from '../../color';
+import { encodePickingColor } from '../color';
 import { a_Color, a_filter, a_Position, a_vertexId } from './commonFeatureFunc';
 import { polygonFillTriangulation as triangulation } from './triangulation';
